fix(compilation): comprehensive resolution of all compilation errors
Systematically fixed compilation errors across entire codebase in 7 phases:

Phase 1: Interface Implementations
- Added missing methods to EmbeddingProvider, BackpressureManager, BulkheadManager,
  CircuitBreakerManager, and RateLimiterManager interfaces
- Fixed method signatures and return types

Phase 2: Duplicate Definitions
- Resolved L1Cache/L2Cache/L3Cache redeclaration by renaming constants
- Fixed duplicate function names in generics package
- Removed duplicate AuthConfig struct

Phase 3: Struct Usage
- Fixed NetworkIntentSpec/Status field references
- Updated PackageReference field names (Name→PackageName)
- Fixed struct literal field mismatches

Phase 4: Missing Imports
- Added fmt, strings, context, sync imports where needed
- Fixed metav1 and corev1 imports
- Added missing external dependencies (websocket, grpc)

Phase 5: Type Issues
- Fixed generic type instantiation for CacheEntry
- Corrected AWS SDK string pointer usage
- Fixed webhook.Server interface confusion
- Fixed cache type definitions

Phase 6: Package-Specific Fixes
- pkg/chaos: Fixed pod phase constants
- pkg/generics: Removed unused variables
- pkg/controllers: Added factory functions
- pkg/monitoring: Added CircularBuffer type
- pkg/testutils: Updated fixtures for current API

Phase 7: Clean Up
- Removed unused imports and variables
- Fixed method call syntax errors
- Corrected atomic operations usage

This ensures the entire codebase compiles without errors, enabling CI to pass.

+++ pkg/rag/optimized_rag_pipeline.go
@@ -211,7 +211,7 @@ func NewOptimizedRAGPipeline(
 	}
 
 	// Initialize embedding cache - use the in-memory implementation
-	embeddingCache := NewInMemoryCache(config.EmbeddingCacheSize)
+	embeddingCache := NewInMemoryCache(int64(config.EmbeddingCacheSize))
 
 	pipeline := &OptimizedRAGPipeline{
 		config:            config,
